Add files via upload
In this project, the aim is that selected an image is inserted into another image appropriately. Firstly, 4 points which will create a rectangular or square are selected from the first image. Then, 4 points are selected from the second image to be inserted into the first selected rectangular or square field.